+++ src/test/java/e2e/NestedModulesTest.java
@@ -42,7 +42,18 @@ public class NestedModulesTest {
 
     @Test
     public void buildsAndInstallsAndTagsAllModules() throws Exception {
-        buildsEachProjectOnceAndOnlyOnce(testProject.mvnRelease("1"));
+        buildsAndInstallsAndTagsAllModules(false);
+    }
+
+    @Test
+    public void buildsAndInstallsAndTagsAllModulesAlsoWhenTaggedAtTheEnd() throws Exception {
+        buildsAndInstallsAndTagsAllModules(true);
+    }
+
+    private void buildsAndInstallsAndTagsAllModules(boolean pushTagsAtTheEnd) throws Exception {
+        buildsEachProjectOnceAndOnlyOnce(testProject.mvnRelease("1",
+                "-DpushTagsAtTheEnd=" + pushTagsAtTheEnd
+        ));
         installsAllModulesIntoTheRepoWithTheBuildNumber();
 
         assertBothReposTagged("nested-project", expectedAggregatorVersion, "1");